Removed all *references* to Name, NameSeries, and Pronoun models; ordered migrations up to that point
Anyone working on the create_gender_model branch up to this commit may have to delete their database if they started with a database that used previous versions of migration files.

Up to now, the Name, NameSeries, and Pronoun models are still present in models.py; the next step is deleting them safely.