Reformatted `dict` update in `generate_gender_token_counters`
I also decided to pull the `Gender` object from the database earlier in the `generate_gender_token_counters` function rather than in `generate_token_counters`. The `Gender` objects are slim enough that we don't have to be so careful about lazy evaluation: in this case, this reduces the need for an extra query to hit the database just to get the `label` field value.

Co-Authored-By: sharabhuiyan <sharabhuiyan725@gmail.com>