Improve `Corpus` model (#40)
This PR focuses on improving the Corpus model by introducing some of the optimizations for Document iteration described in PR #26. It also adds some console and string representation for the Corpus and Document models as well as updating some of their docstrings. Furthermore, this PR includes the id field in the serialization of `PronounSeries` and `Gender` instances.