Implement `Document` iteration optimization in `proximity.py` (#38)
This PR lays the foundation for the web counterpart of `proximity.py` and implements the optimization suggested in PR #26 into that file, which was pasted over from the [Gender Analysis Tookit](https://github.com/dhmit/gender_analysis). Instead of iterating through all Document objects associated with a `Corpus` instance, this fix delays `QuerySet` evaluation until absolutely needed by making use of its lazy evaluation and the `.values_list()` method.

Overall, this PR will serve as the foundation for an option for the user to run a proximity analysis on a `Corpus` with the results displayed on the `Corpora` page.