Create `ProximityAnalysis` model (#50)
This PR features the `ProximityAnalysis` model, which acts as a cache for the results of the analysis functions in `proximity.py`. This PR includes the model, the `admin` registry, a (merged) migration file, and a `ProximitySerializer`, which serializes all fields of the model with the caveats of [using the `str` representation](https://www.django-rest-framework.org/api-guide/relations/#stringrelatedfield) of the related `Corpus` and the [primary key](https://www.django-rest-framework.org/api-guide/relations/#primarykeyrelatedfield) for the `Gender`s.